feat: Add README
I have begun to record progress in the readme file. This is to keep
me up to date with changes in codebas. I am to commit early and often
as this is good practise. I am also to remember that the commit bodies
are to be no more than 72 characters in length. The subject line is to
be no more than 50 characters in length.